requirements.txt
flask
redis
requests
python-dotenv
gunicorn